moved shim.hpp => shim/cufile.hpp
